removed python2 compat; list specific python versions; update ci python setup
